pyproject.toml
[project]
name = "github-app-name-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "rich>=14.1.0",
]
